+++ include/shared_types/ControlSerializer.hpp
@@ -9,7 +9,7 @@ inline void from_json(const json& j, shared_types::Control& x) {
     x.setIsRaining(j.at("rain_status").get<bool>());
     x.setLightPercentage(j.at("light_percentage").get<int>());
     x.setWindPercentage(j.at("wind_percentage").get<int>());
-    x.setHumidityPercentage(j.at("humidity_percentage").get<int>());
+    x.setMoisturePercentage(j.at("moisture_percentage").get<int>());
     x.setTemperature(j.at("temperature").get<float>());
 }
 
@@ -19,7 +19,7 @@ inline void to_json(json& j, const shared_types::Control& x) {
     j["rain_status"]         = x.isRaining();
     j["light_percentage"]    = x.getLightPercentage();
     j["wind_percentage"]     = x.getWindPercentage();
-    j["humidity_percentage"] = x.getHumidityPercentage();
+    j["moisture_percentage"] = x.getMoisturePercentage();
     j["temperature"]         = x.getTemperature();
 }
 } // namespace nlohmann